Clean up code, update IC2 and Mekanism APIs

+++ common/micdoodle8/mods/galacticraft/core/tile/GCCoreTileEntityAirLock.java
@@ -1,7 +1,6 @@
 package micdoodle8.mods.galacticraft.core.tile;
 
 import java.util.ArrayList;
-import codechicken.core.alg.MathHelper;
 import micdoodle8.mods.galacticraft.core.blocks.GCCoreBlocks;
 import net.minecraft.block.Block;
 import net.minecraft.nbt.NBTTagCompound;
@@ -41,12 +40,12 @@ public class GCCoreTileEntityAirLock extends TileEntityAdvanced
                 int x = this.lastProtocol.minX + (this.lastProtocol.maxX - this.lastProtocol.minX) / 2;
                 int y = this.lastProtocol.minY + (this.lastProtocol.maxY - this.lastProtocol.minY) / 2;
                 int z = this.lastProtocol.minZ + (this.lastProtocol.maxZ - this.lastProtocol.minZ) / 2;
-                
+
                 if (this.worldObj.getBlockId(x, y, z) != GCCoreBlocks.airLockSeal.blockID)
                 {
-                    this.worldObj.playSoundEffect((double)x, (double)y, (double)z, "galacticraft.player.openairlock", 1.0F, 1.0F);
+                    this.worldObj.playSoundEffect(x, y, z, "galacticraft.player.openairlock", 1.0F, 1.0F);
                 }
-                
+
                 if (this.protocol.minX != this.protocol.maxX)
                 {
                     for (x = this.protocol.minX + 1; x <= this.protocol.maxX - 1; x++)
@@ -54,17 +53,17 @@ public class GCCoreTileEntityAirLock extends TileEntityAdvanced
                         for (y = this.protocol.minY + 1; y <= this.protocol.maxY - 1; y++)
                         {
                             int id = this.worldObj.getBlockId(x, y, z);
-                            
+
                             if (id != 0 && id != GCCoreBlocks.airLockSeal.blockID)
                             {
                                 Block block = Block.blocksList[id];
-                                
+
                                 if (block != null)
                                 {
                                     block.dropBlockAsItem(this.worldObj, x, y, z, this.worldObj.getBlockMetadata(x, y, z), 0);
                                 }
                             }
-                            
+
                             this.worldObj.setBlock(x, y, this.protocol.minZ, GCCoreBlocks.airLockSeal.blockID, 0, 3);
                         }
                     }
@@ -76,17 +75,17 @@ public class GCCoreTileEntityAirLock extends TileEntityAdvanced
                         for (y = this.protocol.minY + 1; y <= this.protocol.maxY - 1; y++)
                         {
                             int id = this.worldObj.getBlockId(x, y, z);
-                            
+
                             if (id != 0 && id != GCCoreBlocks.airLockSeal.blockID)
                             {
                                 Block block = Block.blocksList[id];
-                                
+
                                 if (block != null)
                                 {
                                     block.dropBlockAsItem(this.worldObj, x, y, z, this.worldObj.getBlockMetadata(x, y, z), 0);
                                 }
                             }
-                            
+
                             this.worldObj.setBlock(this.protocol.minX, y, z, GCCoreBlocks.airLockSeal.blockID, 0, 3);
                         }
                     }
@@ -97,12 +96,12 @@ public class GCCoreTileEntityAirLock extends TileEntityAdvanced
                 int x = this.lastProtocol.minX + (this.lastProtocol.maxX - this.lastProtocol.minX) / 2;
                 int y = this.lastProtocol.minY + (this.lastProtocol.maxY - this.lastProtocol.minY) / 2;
                 int z = this.lastProtocol.minZ + (this.lastProtocol.maxZ - this.lastProtocol.minZ) / 2;
-                
+
                 if (this.worldObj.getBlockId(x, y, z) != 0)
                 {
-                    this.worldObj.playSoundEffect((double)x, (double)y, (double)z, "galacticraft.player.closeairlock", 1.0F, 1.0F);
+                    this.worldObj.playSoundEffect(x, y, z, "galacticraft.player.closeairlock", 1.0F, 1.0F);
                 }
-                
+
                 if (this.lastProtocol.minX != this.lastProtocol.maxX)
                 {
                     for (x = this.lastProtocol.minX + 1; x <= this.lastProtocol.maxX - 1; x++)
@@ -110,17 +109,17 @@ public class GCCoreTileEntityAirLock extends TileEntityAdvanced
                         for (y = this.lastProtocol.minY + 1; y <= this.lastProtocol.maxY - 1; y++)
                         {
                             int id = this.worldObj.getBlockId(x, y, z);
-                            
+
                             if (id != 0 && id != GCCoreBlocks.airLockSeal.blockID)
                             {
                                 Block block = Block.blocksList[id];
-                                
+
                                 if (block != null)
                                 {
                                     block.dropBlockAsItem(this.worldObj, x, y, z, this.worldObj.getBlockMetadata(x, y, z), 0);
                                 }
                             }
-                            
+
                             this.worldObj.setBlockToAir(x, y, this.lastProtocol.minZ);
                         }
                     }
@@ -132,17 +131,17 @@ public class GCCoreTileEntityAirLock extends TileEntityAdvanced
                         for (y = this.lastProtocol.minY + 1; y <= this.lastProtocol.maxY - 1; y++)
                         {
                             int id = this.worldObj.getBlockId(x, y, z);
-                            
+
                             if (id != 0 && id != GCCoreBlocks.airLockSeal.blockID)
                             {
                                 Block block = Block.blocksList[id];
-                                
+
                                 if (block != null)
                                 {
                                     block.dropBlockAsItem(this.worldObj, x, y, z, this.worldObj.getBlockMetadata(x, y, z), 0);
                                 }
                             }
-                            
+
                             this.worldObj.setBlockToAir(this.lastProtocol.minX, y, z);
                         }
                     }
